Combined level and stats classes
